+++ src/Core/Definition/RuleDefinitionPropositionCompiler.php
@@ -12,7 +12,6 @@ namespace Cline\Ruler\Core\Definition;
 use Cline\Ruler\Builder\RuleBuilder;
 use Cline\Ruler\Builder\Variable as BuilderVariable;
 use Cline\Ruler\Builder\VariableProperty;
-use Cline\Ruler\Core\CompatibilityMode;
 use Cline\Ruler\Core\Proposition;
 use Cline\Ruler\Exceptions\InvalidRuleStructureException;
 use Cline\Ruler\Exceptions\UnknownRuleOperatorException;
@@ -39,18 +38,17 @@ final class RuleDefinitionPropositionCompiler
     public static function compile(
         RuleDefinition $definition,
         RuleBuilder $ruleBuilder,
-        CompatibilityMode $compatibilityMode = CompatibilityMode::Strict,
     ): Proposition {
         if ($definition instanceof CombinatorRuleDefinition) {
             $method = 'logical'.ucfirst($definition->combinator->value);
 
             if ($definition->combinator === RuleCombinator::Not) {
-                return $ruleBuilder->{$method}(self::compile($definition->operands[0], $ruleBuilder, $compatibilityMode));
+                return $ruleBuilder->{$method}(self::compile($definition->operands[0], $ruleBuilder));
             }
 
             return $ruleBuilder->{$method}(
                 ...array_map(
-                    fn (RuleDefinition $subRule): Proposition => self::compile($subRule, $ruleBuilder, $compatibilityMode),
+                    fn (RuleDefinition $subRule): Proposition => self::compile($subRule, $ruleBuilder),
                     $definition->operands,
                 ),
             );
@@ -66,7 +64,7 @@ final class RuleDefinitionPropositionCompiler
             $fieldString = $definition->field;
 
             /** @var BuilderVariable $builder */
-            $builder = str_contains($fieldString, '.') && $compatibilityMode === CompatibilityMode::Strict
+            $builder = str_contains($fieldString, '.')
                 ? array_reduce(
                     explode('.', $fieldString),
                     /**